# academe/server/openapi.yaml
openapi: 3.1.0
info:
  title: academe API
  version: "1"
  description: |
    Every error is {"error": {"code", "message", "requestId"}}. The app
    switches on `code`; `message` is for people.
    Access tokens last 15 minutes. When one is rejected with `invalid_token`,
    call /auth/refresh once; if that is rejected too, log the student out.
servers:
  - url: http://localhost:8080
paths:
...
  /delete-account:
    post:
      summary: Web account deletion request from academe.cc/delete-account. Stores the request and emails a confirmation to the address. At most one request per email and five per IP address a day; over the limit it answers the same without storing or emailing.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email, maxLength: 254 }
      responses:
        "200":
          description: The page, saying the request was received.
          content: { text/html: { schema: { type: string } } }
        "422":
          description: The page again, saying the email address isn't valid.
          content: { text/html: { schema: { type: string } } }
  /reset-password:
    get:
      summary: Web page behind the Reset password button in the reset email. Shows a new-password form and an Open in the app link (academe://reset?c=…). Does not use up the link.
      description: "Sets a `__Host-academe-reset` cookie (Secure, HttpOnly, SameSite=Strict, 1 hour) over HTTPS (TLS or `X-Forwarded-Proto: https`), or a plain `academe-reset` cookie (HttpOnly, SameSite=Strict) over local http, and puts a form token in the page, an HMAC of the cookie and the link token. Cache-Control no-store, Referrer-Policy no-referrer."
      parameters:
        - { name: c, in: query, required: true, schema: { type: string, maxLength: 64 }, description: The link token. }
      responses:
        "200": { description: The form., content: { text/html: { schema: { type: string } } } }
        "400": { description: No or malformed link; asks to open the email link again., content: { text/html: { schema: { type: string } } } }
    post:
      summary: Set the new password from the web form. Same rules and effects as the app flow — every session ends and every earlier access token answers 401. The student is not logged in on the web.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [c, t, password, confirm]
              properties:
                c: { type: string, description: The link token. }
                t: { type: string, description: The form token from the page. }
                password: { type: string, minLength: 8, maxLength: 128 }
                confirm: { type: string }
      responses:
        "200": { description: Password changed., content: { text/html: { schema: { type: string } } } }
        "403": { description: Missing cookie or a form token that doesn't match this cookie and link., content: { text/html: { schema: { type: string } } } }
        "410": { description: The link expired, was used, or its tries are used up., content: { text/html: { schema: { type: string } } } }
        "422": { description: The passwords differ or aren't 8 to 128 characters; the link is not used up., content: { text/html: { schema: { type: string } } } }
        "429": { description: Too many tries from this IP address., content: { text/html: { schema: { type: string } } } }
...
  /email/{name}:
    get:
      summary: Images and fonts used by emails (logo.png, pebby-wave.png, pebby-shy.png, baloo2-800.ttf, archivo-700.ttf, wordmark-500.ttf, wordmark-600.ttf). Cached for a week.
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: The file. }
        "404": { description: No such file. }
  /.well-known/assetlinks.json:
    get:
      summary: Android App Links statement for com.academe.flutter, built from ACADEME_ANDROID_CERT_SHA256 (comma-separated SHA-256 certificate fingerprints). 404 while it is unset.
      responses:
        "200":
          description: The statement list.
          content: { application/json: { schema: { type: array, items: { type: object } } } }
        "404": { description: No certificate configured. }
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: The access token from a Session. Any bearer route answers 401 `invalid_token` when the token has expired (15 minutes) or was issued before a password reset, a Google link or an account deletion request; refresh or log in again. Other server instances may accept a revoked token for up to 30 seconds.
...